requirements.txt
fastapi
uvicorn
torch
transformers
google.generativeai
supabase
python-dotenv
Pillow